+++ R/mutate_nse.R
@@ -1,10 +1,10 @@
 
-#' mutate non-standard evaluation interface.
+
+
+
+#' mutate non-standard evaluation interface (deprecated).
 #'
-#' Mutate a data frame by the mutateTerms.  Accepts arbitrary text as
-#' mutateTerms to allow forms such as "Sepal.Length >= 2 * Sepal.Width".
-#' Terms are vectors or lists of the form "lhs := rhs".
-#' Semantics are: terms are evaluated left to right if mutate_nse_split_terms==TRUE (the default).
+#' Mutate a data frame by the mutate terms from \code{...} (deprecated, please use \code{\link{mutate_se}}).
 #'
 #' Note: this method as the default setting \code{mutate_nse_split_terms = TRUE}, which while
 #' safer (avoiding certain known \code{dplyr}/\code{dblyr} issues) can be needlessly expensive
@@ -14,34 +14,26 @@
 #' @seealso \code{\link{mutate_se}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_at}}, \code{\link[wrapr]{:=}}
 #'
 #' @param .data data.frame
-#' @param ... stringified expressions to mutate by.
+#' @param ... expressions to mutate by.
 #' @param mutate_nse_split_terms logical, if TRUE into separate mutates (if FALSE instead, pass all at once to dplyr).
 #' @param mutate_nse_env environment to work in.
+#' @param mutate_nse_printPlan logical, if TRUE print the expression plan
 #' @return .data with altered columns.
-#'
-#' @examples
-#'
-#'
-#' resCol1 <- "Sepal_Long"
-#' ratio <- 2
-#' compCol1 <- "Sepal.Width"
-#'
-#'
-#' datasets::iris %.>%
-#'   mutate_nse(., resCol1 := "Sepal.Length" >= ratio * compCol1,
-#'                 "Petal_Short" := "Petal.Length" <= 3.5) %.>%
-#'   summary(.)
-#'
 #'
 #' @export
 #'
 mutate_nse <- function(.data, ...,
                        mutate_nse_split_terms = TRUE,
-                       mutate_nse_env = parent.frame()) {
+                       mutate_nse_env = parent.frame(),
+                       mutate_nse_printPlan = FALSE) {
+  .Deprecated(new = "mutate_se", old = "mutate_nse")
   # convert char vector into spliceable vector
   # from: https://github.com/tidyverse/rlang/issues/116
   mutateTerms <- substitute(list(...))
-  if(!all(names(mutateTerms) %in% "")) {
+  if(!(is.data.frame(.data) || dplyr::is.tbl(.data))) {
+    stop("seplyr::mutate_nse first argument must be a data.frame or tbl")
+  }
+  if(length(setdiff(names(mutateTerms), ""))>0) {
     stop("seplyr::mutate_nse() all assignments must be of the form a := b, not a = b")
   }
   # mutateTerms is a list of k+1 items, first is "list" the rest are captured expressions
@@ -56,10 +48,12 @@ mutate_nse <- function(.data, ...,
         stop("mutate_nse terms must be of the form: sym := expr")
       }
       lhs[[i-1]] <- as.character(prep_deref(ei[[2]], mutate_nse_env))
-      rhs[[i-1]] <- deparse(prep_deref(ei[[3]], mutate_nse_env))
+      rhs[[i-1]] <- paste(deparse(prep_deref(ei[[3]], mutate_nse_env)), collapse = "\n")
     }
-    res <- mutate_se(res, lhs := rhs, splitTerms = mutate_nse_split_terms,
-                     env=mutate_nse_env)
+    res <- mutate_se(res, lhs := rhs,
+                     splitTerms = mutate_nse_split_terms,
+                     env = mutate_nse_env,
+                     printPlan = mutate_nse_printPlan)
   }
   res
 }